requirements.txt
Flask==2.2.3
Flask_Bootstrap==3.3.7.1
Flask_CKEditor==0.4.6
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
python-dotenv==1.0.0
WTForms==3.0.1
